requirement_list.txt
Django==1.11.20
django-filters==0.2.1
djangorestframework==3.9.2
Markdown==3.0.1